+++ src/atoms/atoms.jl
@@ -1,3 +1,11 @@
+"""
+    Atoms
+
+Atoms objects characterize structural representations of molecular structures.
+These representations can either be standardized forms
+(like graph representation using an adjacency matrix, point cloud representation, etc),
+or be custom user-defined representations too.
+"""
 module Atoms
 
 using ..ChemistryFeaturization.AbstractType: AbstractAtoms